model 17.0 gestational age and stillbirth artifact (#135)
Category: artifact
JIRA issue: https://jira.ihme.washington.edu/browse/MIC-6376
Research reference: gestational age and stillbirth

Changes and notes
Added keys to artifact.

Verification and Testing
Built artifact and checked that 1) data were the same shape as in nutrition optimization if the key was in the NO artifact, or that 2) data were in a shape that could be used by the relevant components in the code we will be porting from NO if the key was not in the NO artifact.